# lib/api-zod/src/generated/types/createApiKeyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AIWatch API - Unified AI update intelligence feed
 * OpenAPI spec version: 1.0.0
 */

export interface CreateApiKeyBody {
  name: string;
}